fixing more errors

+++ Sources/QuickLMDBMacros/MDB_comparable.swift
@@ -61,10 +61,10 @@ internal struct MDB_comparable_macro:MemberMacro, ExtensionMacro {
 						return 0
 					
 					case (_, 0):
-						return lhs!.pointee.mv_size
+						return Int32(lhs!.pointee.mv_size)
 						
 					case (0, _):
-						return -1 * rhs!.pointee.mv_size
+						return Int32(-1 * rhs!.pointee.mv_size)
 					
 					case (MemoryLayout<RAW_staticbuff_storetype>.size, MemoryLayout<RAW_staticbuff_storetype>.size):
 						return Self.RAW_compare(lhs_data:lhs!.pointee.mv_data, lhs_count:lhs!.pointee.mv_size, rhs_data:rhs!.pointee.mv_data, rhs_count:rhs!.pointee.mv_size)
